form with validation working
